Update gem module nesting
