+++ lib/capistrano/db_tasks.rb
@@ -1,6 +1,6 @@
 require_relative 'db_tasks/util'
 require_relative 'db_tasks/database'
-require_relative 'db_tasks/asset'
+require_relative 'db_tasks/assets'
 
 require_relative 'db_tasks/compressors/base'
 require_relative 'db_tasks/compressors/bzip2'
