+++ lib/capistrano/db_tasks/compressors/base.rb
@@ -1,3 +1,7 @@
-module Compressors
-  class Base; end
+module Capistrano
+  module DbTasks
+    module Compressors
+      class Base; end
+    end
+  end
 end
