+++ lib/capistrano/db_tasks/compressors/gzip.rb
@@ -1,36 +1,40 @@
-module Compressors
-  class Gzip < Base
-    class << self
-      def file_extension
-        "gz"
-      end
+module Capistrano
+  module DbTasks
+    module Compressors
+      class Gzip < Base
+        class << self
+          def file_extension
+            "gz"
+          end
 
-      def compress(from, to = nil)
-        from = from == :stdin ? "-" : from
-        to = case to
-             when '-'
-               "-c --stdout"
-             when nil
-               ""
-             else
-               "-c --stdout > #{to}"
-             end
+          def compress(from, to = nil)
+            from = from == :stdin ? "-" : from
+            to = case to
+                 when '-'
+                   "-c --stdout"
+                 when nil
+                   ""
+                 else
+                   "-c --stdout > #{to}"
+                 end
 
-        "gzip #{from} #{to}"
-      end
+            "gzip #{from} #{to}"
+          end
 
-      def decompress(from, to = nil)
-        from = from == :stdin ? "-" : from
-        to = case to
-             when :stdout
-               "-c --stdout"
-             when nil
-               ""
-             else
-               "-c --stdout > #{to}"
-             end
+          def decompress(from, to = nil)
+            from = from == :stdin ? "-" : from
+            to = case to
+                 when :stdout
+                   "-c --stdout"
+                 when nil
+                   ""
+                 else
+                   "-c --stdout > #{to}"
+                 end
 
-        "gzip -d #{from} #{to}"
+            "gzip -d #{from} #{to}"
+          end
+        end
       end
     end
   end
