+++ lib/capistrano/db_tasks/util.rb
@@ -1,6 +1,10 @@
-module Util
-  def self.prompt(msg, prompt = "(y)es, (n)o ")
-    ask(:answer, "#{msg} #{prompt} ? ")
-    (fetch(:answer) =~ /^y$|^yes$/i).to_i.zero?
+module Capistrano
+  module DbTasks
+    module Util
+      def self.prompt(msg, prompt = "(y)es, (n)o ")
+        ask(:answer, "#{msg} #{prompt} ? ")
+        (fetch(:answer) =~ /^y$|^yes$/i).to_i.zero?
+      end
+    end
   end
 end
