+++ lib/capistrano/db_tasks/version.rb
@@ -1,3 +1,5 @@
-module CapistranoDbTasks
-  VERSION = "0.6".freeze
+module Capistrano
+  module DbTasks
+    VERSION = "0.6".freeze
+  end
 end
