+++ lib/capistrano/tasks/app.cap
@@ -4,11 +4,11 @@ namespace :app do
     task sync: "capistrano_db_tasks:check_can_push" do
       if fetch(:skip_data_sync_confirm) || Util.prompt("Are you sure you want to REPLACE THE REMOTE DATABASE AND your remote assets with local database and assets(#{fetch(:assets_dir)})")
         on roles(:db) do
-          Database.local_to_remote(self)
+          Capistrano::DbTasks::Database.local_to_remote(self)
         end
 
         on roles(:app) do
-          Asset.local_to_remote(self)
+          Capistrano::DbTasks::Asset.local_to_remote(self)
         end
       end
     end
@@ -17,15 +17,15 @@ namespace :app do
   namespace :local do
     desc "Synchronize your local assets AND database using remote assets and database"
     task :sync do
-      puts "Local database     : #{Database::Local.new(self).database}"
+      puts "Local database     : #{Capistrano::DbTasks::Database::Local.new(self).database}"
       puts "Assets directories : #{fetch(:local_assets_dir)}"
-      if fetch(:skip_data_sync_confirm) || Util.prompt("Are you sure you want to erase your local database AND your local assets with server database and assets(#{fetch(:assets_dir)})")
+      if fetch(:skip_data_sync_confirm) || Capistrano::DbTasks::Util.prompt("Are you sure you want to erase your local database AND your local assets with server database and assets(#{fetch(:assets_dir)})")
         on roles(:db) do
-          Database.remote_to_local(self)
+          Capistrano::DbTasks::Database.remote_to_local(self)
         end
 
         on roles(:app) do
-          Asset.remote_to_local(self)
+          Capistrano::DbTasks::Asset.remote_to_local(self)
         end
       end
     end
