+++ lib/capistrano/tasks/database.cap
@@ -3,8 +3,8 @@ namespace :db do
     desc "Synchronize your remote database using local database data"
     task sync: "capistrano_db_tasks:check_can_push" do
       on roles(:db) do
-        if fetch(:skip_data_sync_confirm) || Util.prompt("Are you sure you want to REPLACE THE REMOTE DATABASE with local database")
-          Database.local_to_remote(self)
+        if fetch(:skip_data_sync_confirm) || Capistrano::DbTasks::Util.prompt("Are you sure you want to REPLACE THE REMOTE DATABASE with local database")
+          Capistrano::DbTasks::Database.local_to_remote(self)
         end
       end
     end
@@ -14,9 +14,8 @@ namespace :db do
     desc "Synchronize your local database using remote database data"
     task :sync do
       on roles(:db) do
-        puts "Local database: #{Database::Local.new(self).database}"
-        if fetch(:skip_data_sync_confirm) || Util.prompt("Are you sure you want to erase your local database with server database")
-          Database.remote_to_local(self)
+        if fetch(:skip_data_sync_confirm) || Capistrano::DbTasks::Util.prompt("Are you sure you want to erase your local database with server database")
+          Capistrano::DbTasks::Database.remote_to_local(self)
         end
       end
     end
